Modals Improvments (#278)
* feature: One modal base to rule them all

* fix: Types

* feature: Refactored modal to support twitter

* fix: Replaced add fields modal

* feature: Refactored import from linktree modal

* feature Nft modal

* fix: Refactor template modal

* fix: Refactor configure poaps modal

* cleanup: Deleted modal base

* replace: Publish Nimi modal

* fix: Deleted old modal

* fix: Styles cleanup

* fix: renaming and further style cleanup

* fix: Dried up shared footer and header

* feature: Moved all models to a new folder

* fix: Linktree imports

* fix: Imports

* fix: Import order

* fix: Remove unused exports

* feauture: Added animation for import buttons and social link buttons

* fix: Added intial false for ssr
