+++ src/components/Modal/styled.ts
@@ -1,7 +1,6 @@
 import { Property } from 'csstype';
-import styled, { keyframes } from 'styled-components';
+import styled from 'styled-components';
 
-import CloseIcon from '../../assets/svg/close-icon.svg';
 import { NimiSignatureColor } from '../../theme';
 
 /**
@@ -20,124 +19,50 @@ export interface ModalInnerComponentProps {
   padding?: Property.Padding;
 }
 
-export const StyledModalDialog = styled.div`
-  scrollbar-width: none;
-  position: fixed;
-  width: 100%;
-  height: 100%;
-  top: 0;
-  left: 0;
-  margin: auto;
-  z-index: 989;
-`;
-
-export const animation = keyframes`
-  0% {
-    background-color: rgba(11, 11, 19, 0)
-  }
-
-  to {
-    @-moz-document url-prefix() {
-      background-color: rgba(11, 11, 19, .4)
-    }
-
-    -webkit-backdrop-filter:blur(6px);
-    backdrop-filter:blur(6px)
-  }
-`;
-
-export const StyledModalBackdrop = styled.div`
-  backdrop-filter: blur(6px);
-  animation: ${animation} 0.25s ease-in-out;
-  display: flex;
+export const StyledModalHeader = styled.header<{ padding?: string }>`
+  gap: 32px;
   justify-content: center;
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  overflow-y: scroll;
-`;
-
-export interface StyledModalOutterWrapperProps {
-  maxWidth?: Property.MaxWidth;
-}
-
-/**
- * Outter container/component for the modal.
- */
-export const StyledModalOutterWrapper = styled.div<StyledModalOutterWrapperProps>(
-  ({ maxWidth = MODAL_MAX_WIDTH }) => `
-  margin: auto;
-  z-index: 1000;
-  width: 100%;
-  max-width: ${maxWidth};
-  padding-left: 16px;
-  padding-right: 16px;
-`
-);
-
-export const StyledModalInnerWrapper = styled.div`
-  display: flex;
-  flex-direction: column;
-  box-sizing: border-box;
-  background-color: #ffffff;
-  border-radius: 12px;
-  overflow: hidden;
-  margin: auto;
-`;
-
-export const StyledModalHeader = styled.header<ModalInnerComponentProps>(
-  ({ padding = DEFAULT_INNER_COMPONENT_SPACING }) => `
+  text-align: center;
   font-weight: 500;
   font-size: 20px;
   display: flex;
   align-items: center;
   flex-direction: column;
-  padding: 24px 28px 20px;
-  padding: ${padding};
-  position: relative;
-`
-);
+  padding: ${({ padding }) => padding};
+  padding-bottom: 0;
+`;
 
-export const StyledModalFooter = styled.footer<ModalInnerComponentProps>(
-  ({ padding = DEFAULT_INNER_COMPONENT_SPACING }) => `
+StyledModalHeader.defaultProps = {
+  padding: DEFAULT_INNER_COMPONENT_SPACING,
+};
+
+export const StyledModalFooter = styled.footer<{ padding?: string }>`
   font-size: 15px;
   line-height: 20px;
   display: flex;
   align-items: center;
-  justify-content: space-between;
+  justify-content: center;
   color: #6f6e84;
   padding: 0px 28px 20px;
-  padding: ${padding};
-`
-);
-
-export const StyledModalContent = styled.main<ModalInnerComponentProps>(
-  ({ padding = DEFAULT_INNER_COMPONENT_SPACING }) => `
-    padding: ${padding};
-`
-);
-
-export const StyledCloseModalButton = styled(CloseIcon)`
-  position: absolute;
-  top: 40px;
-  right: 40px;
-  cursor: pointer;
+  padding: ${({ padding }) => padding};
+  padding-top: 0;
 `;
+StyledModalFooter.defaultProps = {
+  padding: DEFAULT_INNER_COMPONENT_SPACING,
+};
+
+export const StyledModalContent = styled.main<{ padding?: string }>`
+  padding: ${({ padding }) => padding};
+`;
+StyledModalContent.defaultProps = {
+  padding: DEFAULT_INNER_COMPONENT_SPACING,
+};
 
 /**
  * Modal Title
  */
-export const StyledModalTitle = styled.div`
+export const ModalTitleBig = styled.div`
   font-weight: 700;
   font-size: 38px;
   ${NimiSignatureColor};
-`;
-export const ModalSubTitle = styled.div`
-  color: #7a7696;
-  font-weight: 500;
-  font-size: 16px;
-  line-height: 28px;
-  margin-top: 16px;
 `;
