+++ src/components/NFTSelector/NFTSelector.styled.tsx
@@ -4,7 +4,6 @@ export const AssetListInnerWrapper = styled.div`
   max-height: 500px;
   width: 100%;
   overflow: auto;
-
   /** Layout for 2 columns */
   display: grid;
   grid-template-columns: repeat(2, 1fr);
