+++ src/components/form/Button/ButtonGroup.tsx
@@ -1,8 +1,9 @@
 import { NIMI_CARDS_WIDTH } from '@nimi.io/card/constants';
+import { motion } from 'framer-motion';
 import { PropsWithChildren } from 'react';
 import styled from 'styled-components';
 
-export const StyledButton = styled.button<{ disabled: boolean }>`
+export const StyledButton = styled(motion.button)<{ disabled: boolean }>`
   display: flex;
   height: 42px;
   border-color: transparent;
@@ -29,7 +30,18 @@ interface ButtonGroupProps {
 
 export const ButtonGroup = ({ id, children, onClick, disabled = false }: PropsWithChildren<ButtonGroupProps>) => {
   return (
-    <StyledButton disabled={disabled} id={id} onClick={onClick}>
+    <StyledButton
+      //set because of ssr rendering
+      initial={false}
+      whileHover={{
+        scale: 1.1,
+        transition: { duration: 0.2 },
+      }}
+      whileTap={{ scale: 0.9 }}
+      disabled={disabled}
+      id={id}
+      onClick={onClick}
+    >
       {children}
     </StyledButton>
   );
