+++ src/modals/AddFieldsModal/LinksSection/LinksSection.tsx
@@ -3,9 +3,9 @@ import { NimiLinkType } from '@nimi.io/card/types';
 import { useTranslation } from 'react-i18next';
 import styled from 'styled-components';
 
-import { renderSVG } from '../../../../../utils';
-import { ButtonGroup } from '../../../../form/Button';
-import { StyledFlexList } from '../../../styled';
+import { ButtonGroup } from '../../../components/form/Button';
+import { renderSVG } from '../../../utils';
+import { StyledFlexList } from '../styled';
 
 const SectionWrapper = styled.div`
   margin-bottom: 30px;
