+++ src/modals/ConfigurePOAPsModal/components/AnimatedSection.tsx
