+++ src/modals/ConfigurePOAPsModal/components/POAPs.tsx
