+++ src/modals/ConfigurePOAPsModal/components/ReorderItem.tsx
@@ -3,7 +3,7 @@ import { Reorder, useDragControls } from 'framer-motion';
 import { useState } from 'react';
 import styled from 'styled-components';
 
-import DotsIcon from '../../../../../assets/svg/dots.svg';
+import DotsIcon from '../../../assets/svg/dots.svg';
 import { POAPPlaceholder, StaticPOAP } from './POAPs';
 
 interface ReorderItemProps {
