+++ src/modals/ConfigurePOAPsModal/components/index.ts
