+++ src/modals/ConfigurePOAPsModal/index.ts
