+++ src/modals/NFTSelectorModal/NFTSelectorModal.tsx
@@ -1,21 +1,9 @@
 import { useState } from 'react';
-import { useTranslation } from 'react-i18next';
 import styled from 'styled-components';
 
-import { Button } from '../../../Button';
-import {
-  Modal,
-  Content as ModalContentBase,
-  Footer as ModalFooterBase,
-  Header as ModalHeaderBase,
-  Title as ModalTitle,
-} from '../../../Modal';
-import { NFTAsset, NFTSelector } from '../../../NFTSelector';
-
-const ModalHeader = styled(ModalHeaderBase)`
-  padding-bottom: 0;
-  justify-content: center;
-`;
+import { Button } from '../../components/Button';
+import { Modal, Content as ModalContentBase, Footer as ModalFooterBase } from '../../components/Modal';
+import { NFTAsset, NFTSelector } from '../../components/NFTSelector';
 
 const ModalContent = styled(ModalContentBase)`
   min-height: 200px;
@@ -32,24 +20,16 @@ const ModalContent = styled(ModalContentBase)`
   }
 `;
 
-const ModalFooter = styled(ModalFooterBase)`
-  justify-content: center;
-`;
-
 export interface NFTSelectorModalProps {
   onClose: (asset?: NFTAsset) => void;
   address: string;
 }
 
 export function NFTSelectorModal({ address, onClose }: NFTSelectorModalProps) {
-  const { t } = useTranslation();
   const [selectedAsset, setSelectedAsset] = useState<NFTAsset>();
 
   return (
-    <Modal maxWidth="560px">
-      <ModalHeader>
-        <ModalTitle>Select an NFT</ModalTitle>
-      </ModalHeader>
+    <Modal title={'Select an NFT'} handleCloseModal={() => onClose(selectedAsset)} maxWidth="560px">
       <ModalContent>
         <NFTSelector
           address={address}
@@ -58,9 +38,11 @@ export function NFTSelectorModal({ address, onClose }: NFTSelectorModalProps) {
           }}
         />
       </ModalContent>
-      <ModalFooter>
-        <Button onClick={() => onClose(selectedAsset)}>{t('close')}</Button>
-      </ModalFooter>
+      <ModalFooterBase>
+        <Button style={{ marginTop: '10px' }} onClick={() => onClose(selectedAsset)}>
+          Submit
+        </Button>
+      </ModalFooterBase>
     </Modal>
   );
 }
